@@ -33,7 +33,7 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "ext-pcntl": "*",
         "ext-posix": "*",
         "zlikavac32/php-enum": "^2.0",
@@ -50,7 +50,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^7.5",
-        "phpspec/phpspec": "^5.1",
+        "phpspec/phpspec": "^6.0",
         "symfony/yaml": "^4.2",
         "ext-sockets": "*",
         "ext-json": "*",
